requirements.txt
numpy
pandas==1.5.3
Pillow
fastapi==0.101.1
matplotlib==3.8.0
mysql-connector-python==8.1.0
scikit_learn==1.3.0
scipy==1.11.3
seaborn==0.12.2
SQLAlchemy==2.0.20
ppscore==1.3.0 
anvil_uplink==0.4.2
PyJWT==2.8.0
anvil-app-server==1.10.1
